+++ lib/jasmine/headless/railtie.rb
@@ -1,0 +1,14 @@
+require 'jasmine/headless/task'
+
+module Jasmine
+  module Headless
+    class Railtie < Rails::Railtie
+      rake_tasks do
+        Jasmine::Headless::Task.new do |t|
+          t.colors = true
+        end
+      end
+    end
+  end
+end
+