+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.2.0
 
-* Add a Rake task
+* Add a Rake task and a default task for Rails
 * Global runner configuration via ~/.jasmine-headless-webkit
 * Custom Jasmine reporter for better user feedback
 * Specify the specs to be run, instead of always running them all